r/aws_cloudfront_anycast_ip_list: Serialize acceptance tests (quota is 1).
