+++ internal/service/cloudfront/cloudfront_test.go
@@ -1,0 +1,36 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
+package cloudfront_test
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-provider-aws/internal/acctest"
+	"github.com/hashicorp/terraform-provider-aws/names"
+)
+
+func init() {
+	acctest.RegisterServiceErrorCheckFunc(names.CloudFrontServiceID, testAccErrorCheckSkipFunction)
+}
+
+func testAccErrorCheckSkipFunction(t *testing.T) resource.ErrorCheckFunc {
+	return acctest.ErrorCheckSkipMessagesContaining(t,
+		"InvalidParameterValueException: Unsupported source arn",
+	)
+}
+
+func TestAccCloudFront_serial(t *testing.T) {
+	t.Parallel()
+
+	testCases := map[string]map[string]func(t *testing.T){
+		"AnycastIPList": {
+			acctest.CtBasic:      testAccAnycastIPList_basic,
+			acctest.CtDisappears: testAccAnycastIPList_disappears,
+			"tags":               testAccAnycastIPList_tags,
+		},
+	}
+
+	acctest.RunSerialTests2Levels(t, testCases, 0)
+}
